% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primus.R
\name{calcBIC}
\alias{calcBIC}
\title{Computes bayesian information criterion (BIC) for multiple fits}
\usage{
calcBIC(Y, g, w = NULL, fits = NULL)
}
\arguments{
\item{Y}{Raw counts matrix.}

\item{g}{Size factors for cells in B. (default 1).}

\item{w}{Weight for each cell. This is only useful if you have weighted each cell differently in fitting. (default NULL).}

\item{fits}{A list of fits with different different number of latent clusters (k).}
}
\value{
A vector of BICs for the input fits.
}
\description{
Computes bayesian information criterion (BIC) for multiple fits
}
