% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMUS-package.R
\docType{package}
\name{PRIMUS-package}
\alias{PRIMUS}
\alias{PRIMUS-package}
\title{PRIMUS: Poisson scRNA Integration of Mixed Unknown Signals}
\description{
This package implements the PRIMUS algorithm, a holistic clustering approach that identifies phenotypic cell groups from the scRNA-seq data while accounting for patient-specific components as well as technical noise.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KaiyangZ/PRIMUS}
  \item Report bugs at \url{https://github.com/KaiyangZ/PRIMUS/issues}
}

}
\author{
\strong{Maintainer}: Kaiyang Zhang \email{kaiyang.zhang@helsinki.fi}

}
\keyword{internal}
