% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primus.R
\name{primus_centroid}
\alias{primus_centroid}
\title{Computes a denoised centroid profiles for any subset of the data}
\usage{
primus_centroid(A, B, g = 1, w = 1)
}
\arguments{
\item{A}{Effective rate for the nuisance part (like X \%*\% D) for a subset of cells.}

\item{B}{Raw counts matrix the same subset of cells as in A.}

\item{g}{Size factors for cells in B. (default 1).}

\item{w}{Weight for each cell. This is only useful if you want to weight each cell differently. (default 1).}
}
\value{
A matrix of denoised centroid profiles for the subset.
}
\description{
Computes a denoised centroid profiles for any subset of the data
}
