% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primus.R
\name{runPrimus}
\alias{runPrimus}
\title{Poisson scRNA integration of mixed unknown signals}
\usage{
runPrimus(Y, D, k, g = 1, w = 1, max.iter = 100L, min.delta = 1.48e-09)
}
\arguments{
\item{Y}{Input raw counts matrix. Rows represent genes and columns present cells.}

\item{D}{Design matrix of nuisance factors. To remove data source specific signals, this is an adjacency matrix specifing the data source labels of cells in Y. Rows represent nuisance factors. Columns represent cells, corresponding to those in Y.}

\item{k}{Number of latent phenotypic groups.}

\item{g}{Size factors for cells in Y. (default 1).}

\item{w}{Weight for each cell. This is only useful if you want to weight each cell differently. (default 1).}

\item{max.iter}{Maximum of iterations. (default 100).}

\item{min.delta}{Minimum objective change. Large values allow the optimizer to stop early but may result in less accurate parameters. (default 1.48e-9).}
}
\value{
A list with the following members:
\item{$Z}{ The denoised profiles for each cluster. }
\item{$L}{ The identified denoised cell cluster labels for each cell. }
\item{$X}{ The profiles for each nuisance factor. }
\item{$lik}{ The likelihood of the fit. }
}
\description{
\code{runPrimus} identifies phenotypic cell groups from scRNA-seq data while accounting for nuisance factors (e.g. data source specific signals, technical noises).
}
